Add updated overcommit configuration
We updated the way that we configure our linters so that each repository
now holds its own copy of the config, rather than relying on symlinking
to a different repository. (This allows us to run linters in CI.)

This repo was missed in the transition. We removed the underlying files
from the local repository, so this was left with symlinks pointing at
nothing. This caused overcommit to fail.

This change updates the linter configuration to be self-contained, which
matches our other repos and allows overcommit to execute successfully.