INFRA-2483 remove aws cli from Buildkite agents
This change essentially reverts [INFRA-2408](https://github.com/panorama-ed/buildkite-assets/pull/153)
which installed the AWS CLI in the nested Docker container
that the panorama_command script spins up to run secure
commands. We no longer need this CLI here since we opted
to use the ECR Buildkite addon to push parser images.
This change also includes updating the file in
the buildkite-runners AWS Account's S3 bucket (not
shown in this code change, but will be a follow-up).